+++ crypto/sha.h
@@ -6,6 +6,6 @@
 
 #define SHA256_MD_SZ (256U / 8U) // Message digest size in bytes for SHA-256
 
-int sha256(const uint8_t *msg, uint64_t len, uint8_t *md);
+int sha256(const uint8_t *msg, size_t len, uint8_t *md);
 
 #endif // SHA_H
